fix(client): fix types
